Address iOS ProgressBar wrapper feedback
