+++ src/Core/src/PublicAPI/net-ios/PublicAPI.Unshipped.txt
@@ -2,7 +2,6 @@
 override Microsoft.Maui.Handlers.LabelHandler.GetDesiredSize(double widthConstraint, double heightConstraint) -> Microsoft.Maui.Graphics.Size
 override Microsoft.Maui.Handlers.ProgressBarHandler.NeedsContainer.get -> bool
 override Microsoft.Maui.Handlers.ProgressBarHandler.PlatformArrange(Microsoft.Maui.Graphics.Rect rect) -> void
-override Microsoft.Maui.Handlers.ProgressBarHandler.SetupContainer() -> void
 override Microsoft.Maui.Handlers.ShapeViewHandler.GetDesiredSize(double widthConstraint, double heightConstraint) -> Microsoft.Maui.Graphics.Size
 override Microsoft.Maui.Handlers.StepperHandler.GetDesiredSize(double widthConstraint, double heightConstraint) -> Microsoft.Maui.Graphics.Size
 override Microsoft.Maui.Platform.MauiTextView.TextAlignment.get -> UIKit.UITextAlignment
